requirements.txt
colorama==0.4.4
discord.py==2.2.3
ping3==4.0.4
